Fixed the issue with the XrmLookup control losing search term on click of search button
The XrmLookup control will attemp to find a single record to select when the change event is triggerred.
This even is triggered when focus is lost. i.e. tab or click out of the textbox.
If no match is made, the textbox is cleared.
If, however, you click on the search button, the search results should show.
Because the textbox is cleared, the search button submits an empty search term.
Now the search term is persisted in a hidden control for the search button to access.

I added a hidden control to store the value.
Updating the value in the change event, then I use that value on click.
Now the textbox clears when it needs to and the search works.