Updated To use EarlyBoundGenerator.Api.
Removed Previous CrmSvcUtil Utilities.
Mapped all settings of earlyboundtypes except generateStateEnums and GenerateGlobalOptionSets.   There are no settings currently for this in the EarlyBoundGenerator, and I can't think of a good reason why you would not want to include these.

Updated EarlyBoundTypeConfig to inherit from DLaB.EarlyBoundGeneator.Settings.POCO.ExtensionConfig.  This means all settings in the EarlyBoundGenerator will be allowed (except it they are overriden with the current settings of the EarlyBoundTypeConfig.  Also added names for the action, and optionSet.

Updated Tests.

Updated VS.  Not sure that matters...